refactor: split user data access routes
